+++ build-logic/convention/src/main/kotlin/HiltConventionPlugin.kt
@@ -38,7 +38,7 @@ class HiltConventionPlugin : Plugin<Project> {
                 }
             }
 
-            /** Add support for Android modules, based on [AndroidBasePlugin] */
+            // Add support for Android modules, based on AndroidBasePlugin
             pluginManager.withPlugin("com.android.base") {
                 apply(plugin = "dagger.hilt.android.plugin")
                 dependencies {