+++ build-logic/convention/src/main/kotlin/com/z1nt/buildkit/AndroidCompose.kt
@@ -26,9 +26,7 @@ import org.jetbrains.kotlin.compose.compiler.gradle.ComposeCompilerGradlePluginE
 /**
  * Configure Compose-specific options
  */
-internal fun Project.configureAndroidCompose(
-    commonExtension: CommonExtension,
-) {
+internal fun Project.configureAndroidCompose(commonExtension: CommonExtension) {
     commonExtension.apply {
         buildFeatures.apply {
             compose = true