+++ build-logic/convention/src/main/kotlin/com/z1nt/buildkit/GradleManagedDevices.kt
@@ -24,9 +24,7 @@ import org.gradle.kotlin.dsl.invoke
 /**
  * Configure project for Gradle managed devices
  */
-internal fun configureGradleManagedDevices(
-    commonExtension: CommonExtension,
-) {
+internal fun configureGradleManagedDevices(commonExtension: CommonExtension) {
     val pixel4 = DeviceConfig("Pixel 4", 30, "aosp-atd")
     val pixel6 = DeviceConfig("Pixel 6", 31, "aosp")
     val pixelC = DeviceConfig("Pixel C", 30, "aosp-atd")
@@ -60,7 +58,7 @@ internal fun configureGradleManagedDevices(
 private data class DeviceConfig(
     val device: String,
     val apiLevel: Int,
-    val systemImageSource: String,
+    val systemImageSource: String
 ) {
     val taskName = buildString {
         append(device.lowercase().replace(" ", ""))